+++ components/SignUpPage/Form.tsx
@@ -1,9 +1,9 @@
 import styled from "styled-components";
-import PasswordInput from "./PasswordInput";
-import EmailInput from "./EmailInput";
+import PasswordInput from "./PasswordInput/PasswordInput";
+import EmailInput from "./EmailInput/EmailInput";
 import { useForm } from "react-hook-form";
 import { useAuthForm } from "./Form.hook";
-import ConfirmPasswordInput from "./ConfirmPasswordInput";
+import ConfirmPasswordInput from "./ConfirmPasswordInput/ConfirmPasswordInput";
 
 export const SignUpForm = () => {
     const {
